% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckParameters.R
\name{CheckParameters}
\alias{CheckParameters}
\title{Function to check if there are no duplicates in the parameter file}
\usage{
CheckParameters(file)
}
\arguments{
\item{file}{The name of the file with the parameters that are used in the calibration.}
}
\description{
This function checks the if there are no duplicates in the file with the new parameter values
}
\examples{
\dontrun{
CheckParameters(file)
}
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
