% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copytestdata.R
\name{copytestdata}
\alias{copytestdata}
\title{Function to copy test data to user specified folder}
\usage{
copytestdata(new_folder,package_folder=NULL)
}
\arguments{
\item{new_folder}{Folder to where the data needs to be copied}

\item{package_folder}{Folder where the R package is installed, if this is not specified during installation leave this empty.}
}
\description{
This function copies the test data from the R package to a user-defined folder. This is necesarry since the model does not read the data from R.

The model needs to be run in a seperate folder.

If the specified folder does not exist, the function will create it at the user defined loaction. If the packages are installed on default path, then the package_folder argument can be kept empty. If not, the user has to provide the path where the R package is installed.


The test data is plot 2 from: Leibniz-Zentrum für Agrarlandschaftsforschung (ZALF) e.V.
Freely available: http://dx.doi.org/10.4228/ZALF.1992.167
Mirschel, Wilfried; Wenkel, Karl Otto; Wegehenkel, Martin; Kersebaum, Kurt Christian; Schindler, Uwe( 2010): Comprehensive multivariable field data set for agro-ecosystem modelling from Muencheberg Experimental Stations in 1992 - 1998 ,Leibniz-Zentrum für Agrarlandschaftsforschung(ZALF e.V.).[doi: 10.4228/ZALF.1992.167]
}
\examples{
\dontrun{
 for Windows:
   copytestdata(new_folder="C:/testdata/",package_folder=NULL)

 for Linux:
   copytestdata(new_folder="~/testdata/",package_folder=NULL)
}
}
\author{
JWM Pullens
}
