% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daisycontrol.R
\name{Daisy.control}
\alias{Daisy.control}
\title{Control various aspects of the RDAISY implementation}
\usage{
Daisy.control(sensitivity = FALSE, calib = FALSE, dflt = FALSE,
  costfunction = NULL, obs = NULL, wdDir = NULL, OutDir = NULL,
  interval = NULL, year = NULL, All = NULL, param_sens = NULL,
  p.config = NULL, ind = NULL, p = NULL)
}
\arguments{
\item{sensitivity}{Specify if a sensitivity analylis is needed}

\item{calib}{Specify if Optimization/calibration is needed.}

\item{dflt}{Specify if you run Daisy for default values}

\item{costfunction}{read.optim (needs explination)}

\item{obs}{observation data with date and measurement, same dimension of the simulation}

\item{wdDir}{working environment path}

\item{OutDir}{output path}

\item{interval}{is the interval of the calibration period, in our case the from 1992 and 1993 drainage year.}

\item{year}{if is TRUE than user gets the all output objectives of the years, if it FALSE, it gives the mean of all calculated annual objective.}

\item{All}{If TRUE Giving all performance measure, if FALSE giving the specified performance measure in the script}

\item{param_sens}{sensitive parameters}

\item{p.config}{This argument will provide the makes om the p-config data table. This argument can either be a file or a data table already loaded in the R environment.}

\item{ind}{index}

\item{p}{parameters}
}
\description{
Allow the user to set some characteristics of the functions in RDaisy, such as calibration and sensitivity analysis
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
