% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcost.R
\name{f.cost}
\alias{f.cost}
\title{Cost function}
\usage{
f.cost(RunFile, showLogFile, PathToDaisy, ctrluser = NULL, ctrldaisy)
}
\arguments{
\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{showLogFile}{Either True or False to show the log file. (default = FALSE)}

\item{PathToDaisy}{Specify the path to the Daisy executable}

\item{ctrluser}{Extra arguments that are needed for user designed Cost function (see  \code{\link{User.control}}. Do not add it when only a single simple run is neded)}

\item{ctrldaisy}{Extra arguments that are needed for Optimization/Calibration and Sensitivity analysis (see  \code{\link{Daisy.control}})}
}
\description{
This function updates the parameters, runs the model with the new parameters and calculates the cost/fit.
}
\examples{
\dontrun{
f.cost(p, p.config, RunFile,showLogFile,PathToDaisy,Morris,DEoptim,
dflt,costfunction,obs,wdDir,OutDir,interval,year,All,ind,param_sens,timeout)
}
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
