% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foptim.R
\name{f.optim}
\alias{f.optim}
\title{Optimising function}
\usage{
f.optim(p, f.cost, RunFile, p.config, obs, sim.file, PathToDaisy)
}
\arguments{
\item{p}{Parameter}

\item{f.cost}{User defined cost function}

\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{p.config}{The parameter file}

\item{obs}{Observation against which the simulation needs to be tested.}

\item{sim.file}{Simulation against which the measured data needs to be tested.}

\item{PathToDaisy}{Specify the path to the Daisy executable}
}
\description{
This function combines UpdateParameters, RunDaisy and the f.cost function. This function can be run if an optimization is needed.
}
\examples{
\dontrun{
f.optim (p, f.cost, RunFile, p.config, obs, sim.file,PathToDaisy)
}
}
\author{
JWM Pullens, M Jabloun
}
