% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fupdate.R
\name{f.update}
\alias{f.update}
\title{Function to update the parameters}
\usage{
f.update(file)
}
\arguments{
\item{file}{The name of the parameter that needs to be changed.}
}
\description{
This function runs the Daisy model.
}
\examples{
\dontrun{
f.update <- function(file)
}



}
\author{
JWM Pullens, M Jabloun
}
