% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDaisy.R
\name{runDaisy}
\alias{runDaisy}
\title{Function to run the Daisy model}
\usage{
runDaisy(RunFile, showLogFile = FALSE,
  PathToDaisy = "C:/Program Files/Daisy 5.49/bin/daisy.exe", Morris = FALSE,
  DEoptim = FALSE, costfunction, obs, wdDir, OutDir, interval, year)
}
\arguments{
\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{showLogFile}{Either True or False to show the log file. (default = FALSE)}

\item{PathToDaisy}{Define where Daisy is installed. NOTE: This user should have admin rights to this folder.}

\item{Morris}{Specify if Morris sensitivity analylis is needed. (from package sensitivity)}

\item{DEoptim}{Specify if Deoptimizaltion is needed. (from package DEoptim)}

\item{costfunction}{read.optim (needs explination)}

\item{obs}{observation data with date and measurement, same dimension of the simulation}

\item{wdDir}{working environment path}

\item{OutDir}{output path}

\item{interval}{is the interval of the calibration period, in our case the from 1992 and 1993 drainage year.}

\item{year}{if is TRUE than user gets the all output objectives of the years, if it FALSE, it gives the mean of all calculated annual objective.}
}
\description{
This function runs the Daisy model.
}
\examples{
\dontrun{
runfile="C:/Daisy 5.31/sample/test.dai"
PathToDaisy = "C:/Daisy 5.31/bin/daisy.exe"
runDaisy(RunFile=RunFile, showLogFile = FALSE, PathToDaisy = PathToDaisy)
}
}
\author{
JWM Pullens, M Jabloun
}
