% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParameters.R
\name{updateParameters}
\alias{updateParameters}
\title{update Parameters}
\usage{
updateParameters(p, p.config, v, f, t)
}
\arguments{
\item{p}{The parameter}

\item{p.config}{The new parameter value}

\item{v}{The value to which the defined parameter has to change.}

\item{f}{The file in which the parameter is located.}

\item{t}{The file to which the parameter set has to be written.}
}
\description{
This function reads the parameter settings
}
\examples{
\dontrun{
updateParameters(p= "", p.config="")
}
}
\author{
JWM Pullens, M Jabloun
}
