% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usercontrol.R
\name{User.control}
\alias{User.control}
\title{Control various aspects of the RDAISY implementation}
\usage{
User.control(...)
}
\arguments{
\item{...}{Specify if a extra variable is needed for cost function}
}
\description{
Allow the user to set some characteristics of the functions in RDaisy, such as calibration and sensitivity analysis
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
