% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcost.R
\name{f.cost}
\alias{f.cost}
\title{Cost function}
\usage{
f.cost(p, costfunction, p.config, obs, sim.file, PathToDaisy)
}
\arguments{
\item{p}{Name of the parameter}

\item{costfunction}{Is the cost function. The used cost function is user defined and has to be defined as: f.cost(obs,sim.file)}

\item{p.config}{The parameter file}

\item{obs}{Observation against which the simulation needs to be tested.}

\item{sim.file}{Simulation against which the measured data needs to be tested.}

\item{PathToDaisy}{Specify the path to the Daisy executable}
}
\description{
This function updates the parameters, runs the model with the new parameters and calculates the cost/fit.
}
\examples{
\dontrun{
f.cost(p,costfunction, p.config, obs, sim.file,PathToDaisy)
}
}
\author{
JWM Pullens, M Jabloun
}
