% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foptim.R
\name{f.optim}
\alias{f.optim}
\title{Optimizing parameters}
\usage{
f.optim(p, p.config, obs, sim.file, f.cost)
}
\arguments{
\item{p}{Name of the parameter}

\item{p.config}{The parameter file}

\item{obs}{Observation against which the simulation needs to be tested.}

\item{sim.file}{Simulation against which the measured data needs to be tested.}

\item{f.cost}{Is the cost function. The used cost function is user defined and has to be defined as: f.cost(obs,sim.file)}
}
\description{
This function updates the parameters, rund the model with the new parameters and calculates the cost/fit.
}
\examples{
\dontrun{
updateParameters(p= "", p.config="")
}
}
\author{
JWM Pullens, M Jabloun
}
