% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fupdate.R
\name{f.update}
\alias{f.update}
\title{Function to update the parameters}
\usage{
f.update(p, v, f, t)
}
\arguments{
\item{p}{The name of the parameter that needs to be changed.}

\item{v}{The value to which the defined parameter has to change.}

\item{f}{The file in which the parameter is located.}

\item{t}{The file to which the parameter is stored.}
}
\description{
This function runs the Daisy model.
}
\examples{
\dontrun{
f.update <- function(p, v, f, t)
}

}
\author{
JWM Pullens, M Jabloun
}
