% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fupdate.R
\name{f.update}
\alias{f.update}
\title{Function to update the parameters}
\usage{
f.update(file, RunFile, Morris, DEoptim, dflt, ind)
}
\arguments{
\item{file}{The name of the file with the parameters to be calibrated}

\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{Morris}{Specify if you run Daisy for Morris}

\item{DEoptim}{Specify if you run Daisy for DEOptim}

\item{dflt}{Specify if you run Daisy for default values}

\item{ind}{Index (only use internally for Morris function)}
}
\description{
This function updates the parameter file.
}
\examples{
\dontrun{
f.update(file)
}
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
