% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDaisy.R
\name{runDaisy}
\alias{runDaisy}
\title{Function to run the Daisy model}
\usage{
runDaisy(RunFile, showLogFile = FALSE,
  PathToDaisy = "C:/Program Files/Daisy 5.49/bin/daisy.exe")
}
\arguments{
\item{RunFile}{The complete path to the file with extension ".dai". THis file is thesetup of the model.}

\item{showLogFile}{Either True or False to show the log file. (default = FALSE)}

\item{PathToDaisy}{Define where Daisy is installed. NOTE: This user should have admin rights to this folder.}
}
\description{
This function runs the Daisy model.
}
\examples{
\dontrun{
runDaisy(RunFile="C:/Daisy 5.31/sample/test.dai",showLogFile = FALSE, PathToDaisy = "C:/Daisy 5.31/bin/daisy.exe")
}
}
\author{
JWM Pullens, M Jabloun
}
