% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDaisy.R
\name{runDaisy}
\alias{runDaisy}
\title{Function to run the Daisy model}
\usage{
runDaisy(RunFile, showLogFile = TRUE,
  PathToDaisy = "C:/Program Files/Daisy 5.49/bin/daisy.exe",
  ctrldaisy = Daisy.control())
}
\arguments{
\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{showLogFile}{Either True or False to show the log file. (default = FALSE)}

\item{PathToDaisy}{Define where Daisy is installed. NOTE: This user should have admin rights to this folder.}

\item{ctrldaisy}{Extra arguments that are needed for Optimization/Calibration and Sensitivity analysis (see  \code{\link{Daisy.control}})}
}
\description{
This function runs the Daisy model.
}
\examples{
\dontrun{
runfile="C:/Daisy 5.31/sample/test.dai"
PathToDaisy = "C:/Daisy 5.31/bin/daisy.exe"
runDaisy(RunFile=RunFile, showLogFile = FALSE, PathToDaisy = PathToDaisy)
}
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
