% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParameters.R
\name{updateParameters}
\alias{updateParameters}
\title{update Parameters}
\usage{
updateParameters(p, p.config)
}
\arguments{
\item{p}{The parameter}

\item{p.config}{The new parameter value}
}
\description{
This function reads the parameter settings
}
\examples{
\dontrun{
updateParameters(p= "", p.config="")
}
}
\author{
JWM Pullens, M Jabloun
}
