% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParameters.R
\name{updateParameters}
\alias{updateParameters}
\title{update Parameters}
\usage{
updateParameters(
  p,
  p.config,
  RunFile,
  wdDir,
  OutDir,
  sensitivity,
  calib,
  dflt,
  ind,
  param_sens,
  timeout
)
}
\arguments{
\item{p}{The vector of parameter values that will be used to update the input files}

\item{p.config}{This argument will provide the ma,es om the p-config data table. This argument can either be a file or a data table already loaded in the R environment.}

\item{RunFile}{The complete path to the file with extension ".dai". This file is the setup file of the model.}

\item{wdDir}{working environment path}

\item{OutDir}{output path}

\item{sensitivity}{Specify if you run Daisy for Morris}

\item{calib}{Specify if you run Daisy for DEOptim}

\item{dflt}{Specify if you run Daisy for default values}

\item{ind}{index}

\item{param_sens}{sensitive parameters}

\item{timeout}{time out parameter, if Daisy does not finish on time out, the process will be killed}
}
\description{
This function reads the parameter settings
}
\examples{
\dontrun{
updateParameters(p= "", p.config="")
}
}
\author{
JWM Pullens, M Jabloun, D Nagy
}
